FileUpload

+++ app/assets/javascripts/views/programs/show/files.js
@@ -9,15 +9,18 @@ Scbt.Views.ProgramShowFiles = Backbone.View.extend({
   initialize: function(){
     this.listenTo(this.collection, "sync", this.render)
     this.collection.fetch()
-    this.newFile = new Scbt.Models.ProgramFile
-    this.newFile.set("program_id", this.collection.program.get("id"))
+    this.newProgramFile = new Scbt.Models.ProgramFile
+    this.reader = new FileReader()
+
+    this.newProgramFile.set("program_id", this.collection.program.get("id"))
   },
 
   template: JST["programs/show/files"],
   
   render: function(){
     renderedContent = this.template({
-      files: this.collection
+      files: this.collection,
+      change_permission: this.collection.change_permission
     })
     this.$el.html(renderedContent)
     return this
@@ -28,54 +31,61 @@ Scbt.Views.ProgramShowFiles = Backbone.View.extend({
     this.disableAddFile()
 
     var file = event.currentTarget.files[0]
-    this.newFile.set("file_name", file.name)
-    var reader = new FileReader()
-    reader.onload = function(e) {
-      view.newFile.set("file", this.result)
+    this.newProgramFile.set("file_name", file.name)
+    this.reader.onload = function(e){
+      view.newProgramFile.set("file", this.result)
       view.enableAddFile()
     }
-    if (file){
-      reader.readAsDataURL(file);
-    } 
+    this.read(file)
   },
 
   createFile: function(){
     view = this
-    var newFile = this.newFile
-    newFile.unset("id")
     this.toggleUpload()
     if (this.$('input[type=file]')[0].files[0]){
-      
-      newFile.save({},{
-        success: function(){
-          // alert("File has been added")
-          view.collection.add(newFile)
-          view.render()
-          view.toggleAddFile()
-        }, 
-        error: function(req, res){
-          alert(res.responseJSON || res.responseText)
-          view.toggleAddFile()
-        }
-      })
+      this.saveNewFile()
     } else {
       alert("There is no file to upload")
     }
+  },
+
+  saveNewFile: function(){
+    var view = this
+    this.newProgramFile.unset("id")
+    this.newProgramFile.save({},{
+      success: function(){
+        view.collection.add(view.newProgramFile)
+        view.render()
+        view.toggleAddFile()
+      }, 
+      error: function(req, res){
+        alert(res.responseJSON || res.responseText)
+        view.toggleAddFile()
+      }
+    })
   },
 
   destroyFile: function(event){
     var view = this
     id = $(event.currentTarget).attr("id")
     var programFile = this.collection.get(id)
-    programFile.destroy({
-      success: function(req, res){
-        alert("This file has been deleted")
-        view.render()
-      },
-      error: function(req, res){
-        alert(res.responseJSON || res.responseText)
-      }
-    })
+    if (confirm("Are you sure you want to delete " + programFile.get("file_name"))){
+      programFile.destroy({
+        success: function(req, res){
+          view.render()
+        },
+        error: function(req, res){
+          alert(res.responseJSON || res.responseText)
+        }
+      })
+    }
+  },
+
+  read: function(file){
+    if (file){
+      this.reader.abort()
+      this.reader.readAsDataURL(file)
+    } 
   },
 
   disableAddFile: function(){
